% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_books.R
\name{get_books}
\alias{get_books}
\title{Get book details}
\usage{
get_books(book_id, jscmd = "data")
}
\arguments{
\item{book_id}{a single ISBN, or several passed as a list, max 200}

\item{jscmd}{defaults to 'data', pass 'details' for more detailed response}
}
\value{
tibble containing book details, nested
}
\description{
This function retrieves a variety of details about books using ISBN from
the \href{https://openlibrary.org/dev/docs/api/books}{Open Library API}.
}
