% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ol_get.R
\name{ol_get}
\alias{ol_get}
\title{Basic API call}
\usage{
ol_get(path, query = NULL)
}
\arguments{
\item{path}{the API method to call}

\item{query}{call parameters in the form of a list()}
}
\value{
API object
}
\description{
This function is the base GET call to the \href{https://openlibrary.org/dev/docs/api/books}{Open LIbrary API}.
}
